+++ app/src/lib/build-info.js
@@ -1,2 +1,2 @@
-export const APP_BUILD = 'v1.7.109'
-export const BUILD_DATE = '2026-07-09 13:05'
+export const APP_BUILD = 'v1.7.110'
+export const BUILD_DATE = '2026-07-10 09:40'